pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "groq>=1.0.0",
    "numpy>=2.4.2",
    "pypdf>=6.7.5",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
]
